Replace old ResourcePathManager with ResourceManager and virtual file system
